Move utils.expand_module in its own file in lint
We never used it anywhere else than lint so it belong in lint.
