+++ pylint/utils/__init__.py
@@ -47,14 +47,12 @@ from pylint.utils.file_state import FileState
 from pylint.utils.utils import (
     HAS_ISORT_5,
     IsortDriver,
-    _basename_in_ignore_list_re,
     _check_csv,
     _format_option_value,
     _splitstrip,
     _unquote,
     decoding_stream,
     deprecated_option,
-    expand_modules,
     format_section,
     get_global_option,
     get_module_and_frameid,
@@ -70,14 +68,12 @@ __all__ = [
     "ASTWalker",
     "HAS_ISORT_5",
     "IsortDriver",
-    "_basename_in_ignore_list_re",
     "_check_csv",
     "_format_option_value",
     "_splitstrip",
     "_unquote",
     "decoding_stream",
     "deprecated_option",
-    "expand_modules",
     "FileState",
     "format_section",
     "get_global_option",
